requirements.txt
aniso8601==9.0.1
autopep8==2.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.1
cryptography==41.0.5
flake8==7.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
mando==0.7.1
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
packaging==24.0
pluggy==1.4.0
protobuf==4.21.12
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
radon==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.26
typing_extensions==4.8.0
urllib3==2.2.1
Werkzeug==3.0.1
